requirement.txt
opencv-python==4.5.5.64
dlib==19.22.1
fastapi==0.95.0
uvicorn==0.18.3
numpy==1.23.5
Pillow==9.3.0
